pyproject.toml
[project]
name = "remote-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "mcp-server>=0.1.4",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.34.3",
    "yfinance>=0.2.63",
]
